Adding domain events

+++ src/main/java/com/meep/test/application/VehiclesService.java
@@ -1,10 +1,16 @@
 package com.meep.test.application;
 
-import com.meep.test.domain.*;
-import com.meep.test.domain.vehicle.*;
+import com.meep.test.domain.Filter;
+import com.meep.test.domain.vehicle.Vehicle;
+import com.meep.test.domain.vehicle.VehicleWasAdded;
+import com.meep.test.domain.vehicle.VehicleWasRemoved;
+import com.meep.test.domain.vehicle.VehiclesRepository;
+import com.meep.test.domain.vehicle.VehiclesVendorRepository;
+import com.meep.test.shared.domain.DomainEventPublisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -18,9 +24,12 @@ public class VehiclesService {
 
     private VehiclesRepository vehiclesRepository;
 
-    public VehiclesService(VehiclesVendorRepository vehiclesVendorRepository, VehiclesRepository vehiclesRepository) {
+    private DomainEventPublisher domainEventPublisher;
+
+    public VehiclesService(VehiclesVendorRepository vehiclesVendorRepository, VehiclesRepository vehiclesRepository, DomainEventPublisher domainEventPublisher) {
         this.vehiclesVendorRepository = vehiclesVendorRepository;
         this.vehiclesRepository = vehiclesRepository;
+        this.domainEventPublisher = domainEventPublisher;
     }
 
     public void updateVehicles(Filter filter) {
@@ -33,14 +42,26 @@ public class VehiclesService {
                 .filter(dbVehicles::contains)
                 .collect(Collectors.toSet());
         apiVehicles.removeAll(intersection);
-        vehiclesRepository.save(apiVehicles);
-        for (Vehicle vehicle : apiVehicles) {
-            DomainEventPublisher.getInstance().publish(new VehicleWasAdded(vehicle.getId(), vehicle.getName()));
-        }
+        saveNewVehicles(apiVehicles);
         dbVehicles.removeAll(intersection);
-        vehiclesRepository.remove(dbVehicles);
-        for (Vehicle dbVehicle : dbVehicles) {
-            DomainEventPublisher.getInstance().publish(new VehicleWasRemoved(dbVehicle.getId(), dbVehicle.getName()));
+        removeOldVehicles(dbVehicles);
+    }
+
+    private void removeOldVehicles(Set<Vehicle> dbVehicles) {
+        if (!CollectionUtils.isEmpty(dbVehicles)) {
+            vehiclesRepository.remove(dbVehicles);
+            for (Vehicle dbVehicle : dbVehicles) {
+                domainEventPublisher.publish(new VehicleWasRemoved(dbVehicle.getId(), dbVehicle.getName()));
+            }
+        }
+    }
+
+    private void saveNewVehicles(Set<Vehicle> apiVehicles) {
+        if (!CollectionUtils.isEmpty(apiVehicles)) {
+            vehiclesRepository.save(apiVehicles);
+            for (Vehicle vehicle : apiVehicles) {
+                domainEventPublisher.publish(new VehicleWasAdded(vehicle.getId(), vehicle.getName()));
+            }
         }
     }
